+++ crates/cheatnet/src/cheatcodes/mod.rs
@@ -16,6 +16,7 @@ pub mod mock_call;
 pub mod prank;
 pub mod precalculate_address;
 pub mod roll;
+pub mod spoof;
 pub mod spy_events;
 pub mod warp;