+++ crates/cheatnet/tests/cheatcodes/mod.rs
@@ -1,7 +1,9 @@
+mod deploy;
 mod get_class_hash;
 mod mock_call;
 mod prank;
 mod precalculate_address;
 mod roll;
+mod spoof;
 mod spy_events;
 mod warp;