+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `l1_handler_execute` cheatcode
 - support for `roll`ing/`warp`ing/`prank`ing the constructor logic (precalculate address, prank, assert pranked state in constructor)
 - `spy_events` cheatcode
+- support for printing in contracts
+- `spoof` cheatcode
 
 ### Cast
 
@@ -24,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for custom networks - accounts created on custom networks are saved in `accounts-file` under network's
   chain_id
 - `accounts-file` field in Scarb.toml profile
+- Include the class hash of an account contract in the `accounts-file`
 
 #### Removed